feat: remove `Parse.Cloud.httpRequest`
